Moved include.guide and include.solution outside lab, into general scope.
The possibility of including the sections/notes labeled with professorguide
and/or solution was allowed only in the Laboratory.xsl stylesheet. This made
no sense, since such labeling may occur in any context. Besides, the
stylesheet processing these elements are prepared to ignore those elements
by default. As a consequence, they are now imported in HeadTail, the most
generic stylesheet.

Derived from this change, the variables have been renamed to include a prefix
different from "laboratory".