ADA now detects dependencies in XML files introduced with xincludes.
When processing an XML file, the automatic rules in ADA used to detect only
the dependency on changes of the source file and the stylesheet. But if an
XML file brings the content of another, and this last one is changed, ADA
did not detect any change and therefore the rules were not re-executed.

Detecting the dependencies introduced through the use of xinclude in XML files
is a problem equivalent to traversing a DAG. A bash script has been created
in ADA_HOME/bin that performs this task recursively. It uses a new stylesheet
GetIncludes.xsl to fetch the URLs of the included files. It only works with
relative path includes, but they are the ones that make sense to be used.

The script then traverses recursively the obtained files and returns the
collection of files on which the original source file depends. The
rules in Xsltproc.xml then incorporate the result of such execution as an
additional source dependency before the execution of xsltproc.