Enhanced targets to clean more effectively.
Export would put a set of files in a dir. Now, thanks to ANT that does not
delete a directory unless it is empty, the export target tries to delete
the destination dir (when different from source dir). That way, the chances
of leaving spurious empty dirs behind is reduced (if not completelly eliminated)