Added more effective support for bilingual files in the templates.
The problem appears when a variable containing a string needs to be considered
in multiple languages. The choice has been made to keep the variable with
the regular name and use it whenever no language spec is given. For additional
languages that need to co-exist with the default one, add the language
suffix. In the stylesheet, this translates into manipulating always these
variables with a Choice element, choosing the different languages if the
variable "profile.lang" has the right value. Otherwise, the regular variable
name with no extension is used.