modified:   AntImports/AllTargets.xml
Inserted the invocation to Ppt2PDF.xml. It goes right after DBlatex.

	modified:   AntImports/Copyfiles.xml
Added debug.level support.
	new file:   AntImports/Ppt2PDF.xml
Rule to translate Ppt2PDF in batch. Requires OpenOffice 2.4 or higher and
relies on a script in bin to invoke it (the following file)
	new file:   bin/ada_ppt_to_pdf
Script to translate in batch ppt to pdf using openoffice. It cannot be done
directly in ant because the source file is not given as a single parameter
but needs to be in the middle of an argument. Since ant only allows the
source file to be referred with the element <srcfile/>, it cannot be used
as a value in an attribute:

<arg value="Execute.macro.MStoPDF(<srcfile/>)"/> INCORRECT!!