Strengthened the check for *.files to be non empty (Partially fixed!)
If a regular expression is given in foo.files, the check for this variable
to be non-empty is not correct. The pattern will obviously be non-empty. What
it needs to be really checked is the result of expanding such pattern. That is
if the resulting fileset is non empty. This passes by first catching this
fileset in a property. The condition must be then, this property to be non
empty. The solution has been partially implemented in Inkscape.xml and
TEMPLATE.xml (as a reference for the rest of rule files)
	modified:   Inkscape.xml
	modified:   TEMPLATE.xml