Major revamp of the FAQ. All sections were reviewed except 4 Qs in GIT.
Reviewed the written stuff and added figures and changes to the GIT section
created by JAF (excelent work!).

	modified:   FAQ.xml
	modified:   Params.xml
Needed to touch Params because the https was not working from windows clone.
	modified:   images/.gitignore
	new file:   images/GitAuthorBranches.svg
	new file:   images/GitAuthorCloned.svg
	new file:   images/GitBranches.svg
	new file:   images/GitScenario.svg
All the following PNG files are screen captures.
	new file:   images/GitWindowsInstall.png
	new file:   images/GitWindowsInstall2.png
	new file:   images/GitWindowsInstall3.png
	new file:   images/GitWindowsInstall4.png
	new file:   images/GlobalRule.svg
	modified:   images/Properties.txt
Nuked a spurious rule that was present for figures. They moved to images.
	new file:   images/SubrecursiveExport.svg