Added support for fine grain debug.level control.
Each stylefile has a NAME.debug.level which is initialized to
${ada.debug.level}. The user may change all of them together by changing this
variable, or each of them individually. Typical use scenario is the execution
of a specific rule that needs to be explored. By providing a command line
option such as -Dname.debug.level=warning the messages for that rule are the
only ones (aside from the default ones) that appear on the screen.