Modified parameters for the exam rule. Separated generic vs. in-file.
The customization of an exam has two types of parameters. Those that are
include as parameters for the style-sheet (xsl:param elements) and those that
are disguised in the Docbook as certain elements specifically labeled (for
example <para condition="name">).

The former are used for customizations that might apply to an entire directory containing several exams. Whereas the latter, since it is included in a
per-file basis, applies more to low-level customizations that might be
different from one file to another.

	modified:   XslStyles/Exam.xsl
	modified:   XslStyles/ExamParams.xsl
	modified:   XslStyles/TestQuestions.xsl