Added __bogus__file__name to all the definitions of filesets to avoid the fileset to contain all the files in a dir when the includes value is empty.
ANT has a a very interesting behavior when filesets are defined. You include
a dir and a "includes" which is a comma/space separated list of patterns of
file names to consider. The problem appears when the line

includes="${foo}"

is evaluated when foo is empty. In such case, the list of files is constructed
with ALL the files in the directory. This could be catastrophic, for example
if you ask to delete the files specified in the variable foo. If left empty
the entire directory is nuked.

Another undesired effect is when trying to detect how many files are in
a fileset. If foo is empty, it means the fileset should be empty. With the
ANT default behavior, this condition is not satisfied and you might end up
executing a rule as many times as files in a directory, none of which should
execute.

Conflicts:

	AntImports/Dvi2bitmap.xml
	AntImports/Exam.xml
	AntImports/Inkscape.xml
	AntImports/TestExam.xml