Allow execution of recursive dir even when empty .dir is given
When subrecursive.dst.dir is empty, it used to be the case that the recursion
was skipped. But a scenario in which you want recursive build with no
export might be required. For example, when trying to fetch a subset of the
produced files in the remote directory. By setting .dst.dir to empty and then
using copyfiles, such procedure can be easily applied.
	modified:   AntImports/SubRecursive.xml