Massive first boatload of changes toward improving execution times.
Rewriting most of the rules to reduce property overwriting, file parsing, etc.
The rules now rely more heavily on macros (to avoid full file re-parsing) and
there is a minimum amount of properties computed to decide if a rule should
be considered. So far the reduction in the metrics (props overwritten, files
loaded and parsed) is huge. Remains to be seen the true impact while building
large sites.

SampleCourse needed to be modified because MergeStyles has dissapeared and only
one single variable is required and with an absolute path. The FAQ was modified
to reflect this change.