Major overhaul to implement a truly scalable multilingual functionality.
Rules have change as to rely solely in xsltproc with a multilingual variable.
The system then detects the language that have a corresponding rule in
AntImports and iterates over them.

Also, the mergestyles files, if treated in a multilingual processing, are
profiled with the language. That way, parameters for any language can
coexist with different lang="??" attribute. This profiling phase selects only
the relevant parameters.