Some improvements in formatting RSS items inside XHTMLs.
Now the "rss" tag works in two modes:
- Short mode <rss file="..." mode="short" />
Only the first paragraph of the most recent items is shown.
- Normal mode <rss file="..." />
All the paragraphs of all the items are shown.

Items in the long mode are named as ...#item1, ...#item2, etc.

Short mode also links the titles of the items to the page
generated by the long mode if parameter ada.rss.html.path
has a non-empty value.