Fixed major glitch in Extraant.
Targets were not executing properly because of the location of variable
definitions and invocation of rules in AllTargets. One of the changes is that
the invocations are done now after parsing Properties.txt which was not the
case before. The absence of any invocation of ada.init meant that properties
were not properly defined at the level of AllTargets.xml
	modified:   AntImports/AllTargets.xml
	modified:   AntImports/ExtraAnt.xml