Changes in adado and Xsltproc to avoid letting xsltproc get out to the net to fetch DTDs.
I realized that it was constantly going out for some xhtml
DTDs. I've include them in the DTDs directory together with a catalog.template
to generate a catalog file that takes care of these searches. It is yet
unclear the effect of this change in performance. I was working with a really
slow net, that's how I detected the problem. With a fast connection, the
impact could be irrelevant.