IMPORTANT COMMIT: Xsltproc uses now --path option and by default searches for xslstyles in . and ${ada.home}. This means that ADA style sheets may be imported, included, etc by simply saying <Styledir>/<ADAStylesheet>.xsl.
Since the stylesheets were in XslStyles and this dir name will appear in
stylesheets and Properties.txt, decided to rename it to ADA_Styles. That way
whenever an ADA style sheet is being used, since it needs to be referred to as
ADA_Styles/<stylesheet>.xsl, it'll be selfexplanatory where is the style
comming from.