Refined the behavior when redirecting output to build.out
The file was created always. This was inconsistent if ANT does NOTHING, because
the presence of build.out with a recent date/time stemp was taken in
different context as "something happened".

Now if getdependencies.sh is executed in a dir with no build.out and creates
no output (empty file), it deletes the file to preserve the meaning.