Xsltproc.xml now accepts in xsltproc.style.file accepts stylesheet name that is searched by default in the ADA_Style directory. This is an incremental feature increase that the following commit
  commit 905f9406cd17f2ac0c2a45b085a0d583bc51bd54
  Author: Abelardo Pardo <abelardo.pardo@uc3m.es>
  Date:   Sat Nov 15 19:55:03 2008 +0100

left unexplained. Stylesheets can import ADA stylesheets by simply including
the name, but if only the name is used when assigning a value to
xsltproc.style.file and merge was active, it used to create the wrong filename
and bomb out. This has been fixed. Thus:

Properties.txt now may define xsltproc.style.file only with the file name
of a sheet in the ADA_Styles directory. The rule works properly and takes
the sheet from the ADA directory.