ADA now checks if the actual version is correct for execution.
Three new variables are defined:

ada.exact.version
ada.maximum.version
ada.minimum.version

All three need to comply with the format major.minor.point, where major, minor and point are integers between 1 and 99.

The InitialCheck rule includes now a check if any of these three variables is
not empty, and if so, compares it with the current version.

The comparison is done in a bash script because the lack in ANT of opearations
to compare two integers.