Fixed bug related to empty args when invoking "apply" in Cygwin and Unix.
Empty options when running a command are passed as arguments in Linux but
ignored completely in Cygwin (or rather Windows Java). It seems to be a
problem related on how java.exe is invoked. The main impact is that constructs
such as

arg: --stringparam
arg: blah
arg: ${value}

become incorrect if ${value} is empty. In Unix, the empty arg is passed
correctly, but in Windows, it dissappears and the command line becomes
incorrect. Fixed by creating a variable depending on the presence of
${value}. Tested in windows and Linux. No problem.