First version of the RSS auto-link and auto-guid features.
The link and guid for each RSS item can be autogenerated by adding
the item position in the docbook source file to a given prefix.
These features are deactivated by default. To activate them:
  <xsl:param name="ada.rss.autolink">true</xsl:param>
  <xsl:param name="ada.rss.autoguid">true</xsl:param>

If activated, the prefixes must be provided. For example:
  <xsl:param name="ada.rss.autolink.urlprefix"><xsl:value-of
      select="$ada.publish.url"/>/<xsl:value-of
      select="$ada.rss.html.path" /></xsl:param>
  <xsl:param name="ada.rss.autoguid.guidprefix"><xsl:value-of
      select="$ada.rss.autolink.urlprefix"/>2008_09</xsl:param>

The autolink feature makes sense only if $ada.rss.autolink.urlprefix
points to the URL of a page processed by XhtmlAdaFilter.xsl that
includes all the RSS items in XHTML version. This requirement does no
apply to the autoguid feature, which can be used standalone.