adafindcoursehome is invoked only once per DIRECTORY instead of once per target.
By using a candidate variable that is passed as argument when invoking ant
and subant, the rules now only resort to execute the adafindoradd script
when the candidate variable has a bogus value. If not, it is reused. This is
sort of a one single value cache. When properly set, it avoids executing the
script to get it. If not set, the script is then invoked.