Shuffle (lib/TestShuffle) now shuffles answers leaving the last one untouched.
Shuffle is a java program to shuffle a set of qandadivs within a qandaset. Each qandadiv may have several qandaentries. Both qandadivs and qanaentries are
shuffle. Now it also shuffles the answers within the qandaentries. The procedure
has to quirks, though. It only works if the number of answers is greater than
two and the last answer is always left in its position. This is to allow the
use of "None of the above as valid last answer in a question.