Fixed glitch in getdependencies. Silent termination when processing an incorrect XML turned into empty set of dependencies.
When processing the chain of includes in an XML file, if an incorrect file was
found, the xsltproc processor would fail and the script would terminate with
code 1 which is interpreted inside ANT as correct. Also, the execution of
this script was not flagged as failonerror, and therefore the overall
execution would terminate apparently with no problem.

Exec of getdependencies.sh now terminates the build in case of failure.

Bombs out with the proper error code when the internal xsltproc fails.