Refined the treatment of getdependencies.sh output to fix a bug in dependendies.
When using <exec> and catching the output in a property and appending the
error to a file, ant does not behave as it would be expected and the output
is empty. By changing the directive to log the error (instead of appending it
to a file), output catches again the right value.