Fixed a bug when the xml:base attribute in xi:include does not terminate in /
Xsltproc does not choke with this, and if xml:base does not terminate in / it
adds it to concatenate with the href attribute. However, the getdependencies
script did not do such thing. This meant that an attribute such as
xml:base="." would break the dependencies mechanism. The style to detect these
files in an XML file and the script to process the dependencies have been
patched.