Moved error messages in getdependencies to build.out file to simplify visibility.
The build.out file is supposed to collect errors of everything that executes
within ADA. So far the script was simply dumping through stderr some errors
when encountered. Now, the messages are left in buidl.out and the return status
is still -1, thus the error is detected AND SHOWN in the proper location.