Important: Filter now searches files also in ADA_Styles
The dependencies were only applied to XML files, but not to the style files.
When extending the script to consider these files as well, an issue appeared
because xsltproc is invoked with --paths including a set of directories to
search for the style files by default. When searching for dependencies, the
script must mimic this behavior.