Reorganizing dependency detection.
Dependency analysis is no needed in the mergestyle step, because there is only
a simple concatenation happening. However, when regular xsltproc is applied
the dependency must be applied both to the source XML file AND the
stylesheet file. So far, this last set of dependencies were approximated by
simply including all the style files in ADA_Styles, but when private style
files are used and these files have imports or includes, the dependency
mechanism would not detect them.

Conclusion, when merging two stylesheets, the mergestyles step does not need
to use the dependency checking, but when regular xsltproc is applied, the
dependencies are now computed for both source and style file.

Getting dependencies from style files required the filter to detect them
in the XML files to ignore hrefs starting by "http" (previous commit).