Relaxed checking for ada.course.home defined to allow empty string as valid.
Up to here, when ada.course.home was the empty string, it was considered
in InitialCheck as not defined. However, due to the changes introduced
in 9420a7483e835e2cc59343fdd4faef85b57a21ce, adafindcoursehome returns the
empty string when the course home is not found. The condition, then to
invoke adafindcoursehome has been relaxed. When the string is empty, it is
assumed that the course.home has been defined.