Refined div classes for the case of single passwd vs all passwd.
There is a different class for the password box when it is the only one to
type. The problem was that the CSS style needed to be different for the
case when multiple passwords are introduced. The changes in the CSS file
are to render with different styles these elements.