Variable "languages" has now been pushed all the way down to the DEFAULTS.
The reason is because it is common to several rules, thus, it is convenient to
control its value from the DEFAULTS section. Although this was possible until
now, it was due to a bug. When the bug was fixed, the DEFAULT value of
languages was overwritten by all the basic rules, thus rendering the
functionality useless.

As a rule for further study, whenever an option is present in more than one
rule, it should be pushed to the defaults to facilitate control it for all
targets at the same time.