Implemented the hierarchically defined sections/options. Example:
[DEFAULT]
name = valueD

[a]
name = valueA

[a.b]
othername = othervalue

Value of a.b.name?  Two options.

 1.- Behavior of ConfigParser returns valueD because it is in the defaults.

 2.- ADA forced behavior. Sections are hierarchically organized based on the 
     dot-separated section names. Thus, names in section "a" should take
     precedence over defaults. Return valueA.

The changes in this commit implements policy 2 with hierarchical organization of
the sections. It is based on storing all the given definitions in a specially
named section in ConfigParser. When getting the value of an option, the special
section is checked first to see which section is the appropriate to check for
the answer.