Major adjustment in style resolver due to windows compatibility.
Paths may come now in a URL of the style

file:///d:/a/b/c

When removing the file:// prefix we end up with the path

/d:/a/b/c

and when probing if that file exists, the result is always negative. The trick
is to normalize the path, which in windows turns slash in backslashes and the
path is then

\d:\a\b\c

Then check if the first character is '\', if so, we are in a Windows platform
and we need to remove the first slash to have a correct path:

d:\a\b\c

which is properly detected.