Added support for 'dvipdf' as outputFormat in the latex rule.
The 'pdf' output format of Latex is not equivalent to generating a 'dvi'
and then processing it with 'dvipdf'. Some source files do not work
properly with the 'pdf' output format.

This commit adds a third value for ouputFormat: 'dvipdf'. When
specifying this value, the rule generates a DVI file first, and then
executes 'dvipdf' on it to generate the PDF file.