Detection of dependencies in Latex files.
Documents imported from LaTeX files with \input or \include are
now detected as dependencies. Graphic files imported with
\includegraphics are detected as well.

The only non-straightforward change is the 'baseDir' parameter in
addNode() and update(). The baseDir parameter was added because included
relative paths in LaTeX are relative to the root file, whereas in
Docbook they are relative to the location of Docbook file they are
included from. That is, if A includes B and B includes C, in LaTeX C is
included from B with a path relative to the location of A, whereas in
Docbook it is included with a path relative to the location of B.