Running Latex in non-stop mode.
Before this change, Latex entered the interactive mode whenever an error
ocurred in the compilation of the document. As Adagio blocks the output
of Latex, the user perceived the program as blocked.

With this change, Latex never enters interactive mode.

There is still a drawback: errors in the document are not seen in the
terminal, and the user does not know they happened unless she carefully
reviews the output document or reads the adagio.log file. In the future
it would be interesting to parse the output of Latex in order to make
error messages more visible for the user.